Move legal route into ept-legal
