+++ modules/ept-legal/index.js
@@ -1,0 +1,7 @@
+var path = require('path');
+var routes = require(path.normalize(__dirname + '/routes'));
+
+module.exports =  {
+  name: 'legal',
+  routes: routes
+};
