+++ modules/ept-legal/routes/index.js
@@ -1,0 +1,5 @@
+var path = require('path');
+
+module.exports = [
+  require(path.normalize(__dirname + '/legal')),
+];
